+++ src/Stratis.Bitcoin.Features.SmartContracts.Tests/MethodParameterByteSerializerTests.cs
@@ -3,25 +3,27 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using NBitcoin;
+using Stratis.Bitcoin.Features.SmartContracts.Networks;
 using Stratis.SmartContracts;
 using Stratis.SmartContracts.Executor.Reflection.Serialization;
 using Xunit;
 
 namespace Stratis.Bitcoin.Features.SmartContracts.Tests
 {
-    public class MethodParameterStringSerializerTests
+    public class MethodParameterByteSerializerTests
     {
-        public IMethodParameterSerializer Serializer = new MethodParameterStringSerializer();
+        public static Network Network = new SmartContractPosRegTest();
+        public IMethodParameterSerializer Serializer = new MethodParameterByteSerializer(new ContractPrimitiveSerializer(Network));
 
         [Theory]
         [MemberData(nameof(GetData), parameters: 1)]
         public void Roundtrip_Method_Param_Successfully(object value)
         {
             // Roundtrip serialization
-            var methodParamObjects = this.Serializer.Deserialize(this.Serializer.Serialize(new [] { value }));
+            var methodParamObjects = this.Serializer.Deserialize(this.Serializer.Serialize(new[] { value }));
 
             Type paramType = value.GetType();
-                
+
             // Equality comparison using .Equal is possible for these Types
             if (paramType.IsValueType || paramType == typeof(uint160) || paramType == typeof(string))
             {
@@ -31,26 +33,23 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Tests
             // For byte arrays we must compare each element
             if (paramType.IsArray && paramType.GetElementType() == typeof(byte))
             {
-                Assert.True(((byte[])value).SequenceEqual((byte[]) methodParamObjects[0]));
+                Assert.True(((byte[])value).SequenceEqual((byte[])methodParamObjects[0]));
             }
         }
 
         [Fact]
-        public void Serialize_Multiple_Params()
+        public void Roundtrip_Serialize_Multiple_Params()
         {
-            object[] methodParameters =
-            {
-                (int) 12,
-                true,
-                "te|s|t",
-                "te#st",
-                "#4#te#st#",
-                '#'
-            };
+            object[] methodParameters = GetData(0).SelectMany(o => o).ToArray();
 
             var serialized = this.Serializer.Serialize(methodParameters);
 
-            Assert.True(Encoding.UTF8.GetBytes("6#12|1#True|4#te\\|s\\|t|4#te\\#st|4#\\#4\\#te\\#st\\#|3#\\#").SequenceEqual(serialized));
+            var deserialized = this.Serializer.Deserialize(serialized);
+
+            for (var i = 0; i < deserialized.Length; i++)
+            {
+                Assert.Equal(methodParameters[i], deserialized[i]);
+            }
         }
 
         public static IEnumerable<object[]> GetData(int numTests)
@@ -63,7 +62,7 @@ namespace Stratis.Bitcoin.Features.SmartContracts.Tests
             yield return new object[] { (uint)36 }; // MethodParameterDataType.UInt
             yield return new object[] { new uint160(new byte[] { 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1 }) }; // MethodParameterDataType.UInt160
             yield return new object[] { (ulong)29 }; // MethodParameterDataType.ULong
-            yield return new object[] { new Address("0x95D34980095380851902ccd9A1Fb4C813C2cb639") }; // MethodParameterDataType.Address
+            yield return new object[] { new Address("SeMvVcDKTLBrxVua5GXmdF8qBYTbJZt4NJ") }; // MethodParameterDataType.Address
             yield return new object[] { (long)12312321 }; // MethodParameterDataType.Long,
             yield return new object[] { (int)10000000 };// MethodParameterDataType.Int
         }
