+++ src/Stratis.SmartContracts.Executor.Reflection/Serialization/MethodParameterStringSerializer.cs
@@ -4,187 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 using NBitcoin;
-using Nethereum.RLP;
 
 namespace Stratis.SmartContracts.Executor.Reflection.Serialization
 {
-    public class Prefix
-    {
-        public byte Value { get; }
-
-        public Prefix(byte value)
-        {
-            this.Value = value;
-        }
-
-        public Type Type => this.GetType(this.Value);
-
-        public int Length { get; } = 1;
-
-        public Type GetType(byte b)
-        {
-            switch ((MethodParameterDataType)b)
-            {
-                case MethodParameterDataType.Address:
-                    return typeof(Address);
-                case MethodParameterDataType.Bool:
-                    return typeof(bool);
-                case MethodParameterDataType.Byte:
-                    return typeof(byte);
-                case MethodParameterDataType.Char:
-                    return typeof(char);
-                case MethodParameterDataType.String:
-                    return typeof(string);
-                case MethodParameterDataType.Int:
-                    return typeof(int);
-                case MethodParameterDataType.UInt:
-                    return typeof(uint);
-                case MethodParameterDataType.Long:
-                    return typeof(long);
-                case MethodParameterDataType.ULong:
-                    return typeof(ulong);
-                case MethodParameterDataType.UInt160:
-                    return typeof(uint160);
-                case MethodParameterDataType.ByteArray:
-                    return typeof(byte[]);
-                default:
-                    throw new ArgumentOutOfRangeException(nameof(b), b, "Unsupported type");
-            }
-        }
-
-        public static Prefix ForObject(object o)
-        {
-            var type = (byte) GetPrimitiveType(o);
-            return new Prefix(type);
-        }
-
-        private static MethodParameterDataType GetPrimitiveType(object o)
-        {
-            if (o is bool)
-                return MethodParameterDataType.Bool;
-
-            if (o is byte)
-                return MethodParameterDataType.Byte;
-
-            if (o is byte[])
-                return MethodParameterDataType.ByteArray;
-
-            if (o is char)
-                return MethodParameterDataType.Char;
-
-            if (o is string)
-                return MethodParameterDataType.String;
-
-            if (o is uint)
-                return MethodParameterDataType.UInt;
-
-            if (o is uint160)
-                return MethodParameterDataType.UInt160;
-
-            if (o is ulong)
-                return MethodParameterDataType.ULong;
-
-            if (o is Address)
-                return MethodParameterDataType.Address;
-
-            if (o is long)
-                return MethodParameterDataType.Long;
-
-            if (o is int)
-                return MethodParameterDataType.Int;
-
-            // Any other types are not supported.
-            throw new ArgumentOutOfRangeException(nameof(o), o, string.Format("{0} is not supported.", o.GetType().Name));
-        }
-
-        public void CopyTo(byte[] result)
-        {
-            result[0] = this.Value;
-        }
-    }
-
-    public class MethodParameterByteSerializer : IMethodParameterSerializer
-    {
-        private readonly IContractPrimitiveSerializer primitiveSerializer;
-
-        public MethodParameterByteSerializer(IContractPrimitiveSerializer primitiveSerializer)
-        {
-            this.primitiveSerializer = primitiveSerializer;
-        }
-
-        public byte[] Serialize(object[] methodParameters)
-        {
-            if (methodParameters == null)
-                throw new ArgumentNullException(nameof(methodParameters));
-
-            var result = new List<byte[]>();
-
-            foreach (var param in methodParameters)
-            {
-                var encoded = this.Encode(param);
-
-                result.Add(encoded);
-            }
-
-            return RLP.EncodeList(result.Select(RLP.EncodeElement).ToArray());
-        }
-
-        public object[] Deserialize(string[] parameters)
-        {
-            throw new NotImplementedException();
-        }
-
-        public object[] Deserialize(byte[] bytes)
-        {
-            RLPCollection list = RLP.Decode(bytes);
-
-            RLPCollection innerList = (RLPCollection)list[0];
-
-            IList<byte[]> encodedParamBytes = innerList.Select(x => x.RLPData).ToList();
-
-            var results = new List<object>();
-
-            foreach (var encodedParam in encodedParamBytes)
-            {
-                object result = this.Decode(encodedParam);
-
-                results.Add(result);
-            }
-            
-            return results.ToArray();
-        }
-
-        private byte[] Encode(object o)
-        {
-            Prefix prefix = Prefix.ForObject(o);
-
-            byte[] serializedBytes = this.primitiveSerializer.Serialize(o);
-
-            var result = new byte[prefix.Length + serializedBytes.Length];
-
-            prefix.CopyTo(result);
-
-            serializedBytes.CopyTo(result, prefix.Length);
-
-            return result;
-        }
-
-        private object Decode(byte[] bytes)
-        {
-            if (bytes == null)
-                throw new ArgumentNullException(nameof(bytes));
-
-            if (bytes.Length == 0)
-                throw new ArgumentOutOfRangeException(nameof(bytes));
-
-            var prefix = new Prefix(bytes[0]);
-
-            byte[] paramBytes = bytes.Skip(prefix.Length).ToArray();
-
-            return this.primitiveSerializer.Deserialize(prefix.Type, paramBytes);
-        }
-    }
-
     /// <summary>
     /// Class that handles method parameter serialization.
     /// </summary>
